Increase hole size to fit M3
The current M2.5 holes work well for the soldered nuts from JAE
(SM3ZS067U310-NUT1-R1800). However, we are going to use a custom
standoff with a M2 thread (nut) and a screw to mechanically secure the
NVMe. The nut will have a border to fit the hole. That the standoff
still can fit a M2 thread we need M3 holes.

Signed-off-by: Stefan Agner <stefan@agner.ch>