Use impedance for JLC7628 stackup/switch HDMI/SW2 placement
Use JLC7628 impedance values to get more options in color selection.
Also this leads to slightly wider traces typically, and it will be
easier to switch back to JLC2313 or the like than the other way around.

Also swap placing of HDMI/SW2 and rotate the heatsink by 90° for easier
routing.

Signed-off-by: Stefan Agner <stefan@agner.ch>